+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ## 7.0.0
 
 ### Features
+- Add Catalan and Polish support to `detectLanguage`.
 - You can now specify a `hybrid_index` when creating an index for the document index to use hybrid (semantic and keyword) search.
 - `min_score` and `max_results` are now optional parameters in `DocumentIndexClient.SearchQuery`.
 - `k` is now an optional parameter in `DocumentIndexRetriever`.