+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Features
 - Add Catalan and Polish support to `DetectLanguage`.
 - Add utility function `run_is_already_computed` to `Runner` to check if a run with the given metadata has already been computed.
+  - The `parameter_optimization` notebook describes how to use the `run_is_already_computed` function.
 ### Fixes
 ...
 ### Deprecations